+++ src/output/mod.rs
@@ -697,5 +697,4 @@ pub(crate) mod message;
 pub(crate) mod post_exit;
 pub(crate) mod report;
 pub(crate) mod startup_verbose;
-pub(crate) mod ux;
 pub(crate) mod verbose;